version bump and typechecking magic
